+++ src/Plugin/Concerns/HasFormElementTypes.php
@@ -4,7 +4,8 @@ declare(strict_types=1);
 
 namespace CraftCms\Cms\Plugin\Concerns;
 
-use CraftCms\Cms\Cp\FormDefinitions\Elements\FormElement;
+use CraftCms\Cms\Cp\Components\ViewComponent;
+use CraftCms\Cms\Cp\FormDefinitions\Contracts\ProjectableFormElement;
 use CraftCms\Cms\Cp\FormDefinitions\FormElementTypes;
 use CraftCms\Cms\Plugin\Plugin;
 
@@ -13,7 +14,7 @@ use CraftCms\Cms\Plugin\Plugin;
  */
 trait HasFormElementTypes
 {
-    /** @param class-string<FormElement> ...$classes */
+    /** @param class-string<ViewComponent&ProjectableFormElement> ...$classes */
     public function registerFormElementTypes(string ...$classes): void
     {
         $this->app->make(FormElementTypes::class)->registerForPlugin($this, ...$classes);